# docs/api/admin-openapi.yaml
openapi: 3.0.3
info:
  title: tukuyomi Admin API
  version: 1.0.0
  description: |
    OpenAPI document for the admin endpoints exposed by tukuyomi.

    The default admin API base path is `/tukuyomi-api`. It can be changed by configuration,
    so treat these paths as the default contract.
servers:
  - url: /
security: &AdminReadSecurity
  - ApiKeyAuth: []
  - AdminSessionCookie: []
...
paths:
...
  /tukuyomi-api/rules:
    get:
      tags: [waf]
      summary: Get active editable WAF rule assets
      responses:
        '200':
          description: Rule asset map plus per-asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RulesFilesResponse'
    put:
      tags: [waf]
      summary: Save one editable WAF rule file and hot-reload
      security: *AdminWriteSecurity
      parameters:
        - $ref: '#/components/parameters/IfMatchHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleFilePutRequest'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
  /tukuyomi-api/rules:validate:
    post:
      tags: [waf]
      summary: Validate one editable WAF rule file without saving
      security: *AdminWriteSecurity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleFilePutRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /tukuyomi-api/crs-rule-sets:
    get:
      tags: [crs]
      summary: Get CRS rule-set selection state
      responses:
        '200':
          description: CRS selection snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericConfigResponse'
    put:
      tags: [crs]
      summary: Save CRS rule-set selection and hot-reload
      security: *AdminWriteSecurity
      parameters:
        - $ref: '#/components/parameters/IfMatchHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CRSRuleSetsRequest'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '500':
          $ref: '#/components/responses/InternalError'
  /tukuyomi-api/crs-rule-sets:validate:
    post:
      tags: [crs]
      summary: Validate CRS rule-set selection without saving
      security: *AdminWriteSecurity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CRSRuleSetsRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: tukuyomi_admin_session
      description: Signed browser admin session cookie issued by `/tukuyomi-api/auth/login`.
    AdminCSRFCookie:
      type: apiKey
      in: cookie
      name: tukuyomi_admin_csrf
      description: CSRF cookie paired with `tukuyomi_admin_session`.
    AdminCSRFHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF header required for cookie-authenticated non-safe requests.
  parameters:
    IfMatchHeader:
      name: If-Match
      in: header
      required: false
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Optimistic concurrency conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictResponse'
...
    Unprocessable:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationResult'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      additionalProperties: true
    ConflictResponse:
      type: object
      properties:
        error:
          type: string
        currentETag:
          type: string
      additionalProperties: true
    ValidationResult:
      type: object
      properties:
        ok:
          type: boolean
        messages:
          type: array
          items:
            type: string
      additionalProperties: true
    UpdateResult:
      type: object
      properties:
        ok:
          type: boolean
        etag:
          type: string
      additionalProperties: true
...
    RulesFilesResponse:
      type: object
      properties:
        rules:
          type: object
          additionalProperties:
            type: string
        files:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              raw:
                type: string
              etag:
                type: string
              error:
                type: string
            additionalProperties: true
      additionalProperties: true
    GenericConfigResponse:
      type: object
      properties:
        etag:
          type: string
        raw:
          type: string
      additionalProperties: true
...
    RuleFilePutRequest:
      type: object
      required: [path, raw]
      properties:
        path:
          type: string
        raw:
          type: string
...
    CRSRuleSetsRequest:
      type: object
      properties:
        enabled:
          type: array
          items:
            type: string
      additionalProperties: false
...